+++ ms-java-user-management-pecus/src/main/java/de/pecus/api/vo/product/FindDetailProductResponseVO.java
@@ -21,15 +21,15 @@ public class FindDetailProductResponseVO {
 	private String descripcion;
 
 	// Brand
-	private Integer idBrand;
+	private Long idBrand;
 	private String nameBrand;
 
 	//Category
-	private Integer idCategory;
+	private Long idCategory;
 	private String nameCategory;
 
 	//SubCategory
-	private Integer idSubCategory;
+	private Long idSubCategory;
 	private String nameSubCategory;
 
 	public Long getId() {
@@ -49,14 +49,14 @@ public class FindDetailProductResponseVO {
 	public String getDescripcion() {return descripcion;}
 	public void setDescripcion(String descripcion) {this.descripcion = descripcion;}
 
-	public Integer getIdBrand() {return this.idBrand;}
-	public void setIdBrand(Integer idBrand) {this.idBrand = idBrand;}
+	public Long getIdBrand() {return this.idBrand;}
+	public void setIdBrand(Long idBrand) {this.idBrand = idBrand;}
 
-	public Integer getIdCategory() {return this.idCategory;}
-	public void setIdCategory(Integer idCategory) {this.idCategory = idCategory;}
+	public Long getIdCategory() {return this.idCategory;}
+	public void setIdCategory(Long idCategory) {this.idCategory = idCategory;}
 
-	public Integer getIdSubCategory() {return this.idSubCategory;}
-	public void setIdSubCategory(Integer idSubCategory) {this.idSubCategory = idSubCategory;}
+	public Long getIdSubCategory() {return this.idSubCategory;}
+	public void setIdSubCategory(Long idSubCategory) {this.idSubCategory = idSubCategory;}
 
 	public String getNameBrand() {return this.nameBrand;}
 	public void setNameBrand(String nameBrand) {this.nameBrand = nameBrand;}
