+++ ms-java-user-management-pecus/src/main/java/de/pecus/api/vo/product/UpdateProductRequestVO.java
@@ -20,13 +20,13 @@ public class UpdateProductRequestVO {
 		private String descripcion;
 
 		// Product Id Brand
-		private Integer idBrand;
+		private Long idBrand;
 
 		// Product Id Category
-		private Integer idCategory;
+		private Long idCategory;
 
 		// Product Id Subcategory
-		private Integer idSubCategory;
+		private Long idSubCategory;
 
 		public Long getId() {
 			return id;
@@ -49,13 +49,13 @@ public class UpdateProductRequestVO {
 			this.descripcion = descripcion;
 		}
 
-		public Integer getIdBrand() { return idBrand;}
-		public void setIdBrand(Integer idBrand) { this.idBrand = idBrand;}
+		public Long getIdBrand() { return idBrand;}
+		public void setIdBrand(Long idBrand) { this.idBrand = idBrand;}
 
-		public Integer getIdCategory() {return idCategory;}
+		public Long getIdCategory() {return idCategory;}
 		public void setIdCategory() {this.idCategory = idCategory;}
 
-		public Integer getIdSubCategory() {return idSubCategory;}
+		public Long getIdSubCategory() {return idSubCategory;}
 		public void setIdSubCategory() {this.idSubCategory = idSubCategory;}
 
 }
